Rename composer position resolution module
